+++ assets/gulp/src/tasks/unit.ts
@@ -19,7 +19,7 @@ gulp.task("unit-clean", async () => {
     const uniteConfig = await uc.getUniteConfig();
 
     const toClean = [
-        path.join(uniteConfig.dirs.www.unitTestDist, "**/*.spec.js"),
+        path.join(uniteConfig.dirs.www.unitTestDist, "**/*"),
         path.join(uniteConfig.dirs.www.reports, "unit/**/*"),
         path.join(uniteConfig.dirs.www.reports, "coverage/**/*"),
         path.join(uniteConfig.dirs.www.reports, "coverage.json"),